Fix scanner type

+++ src/fr/fabien_jrt/mortgage_calculator/Main.java
@@ -50,7 +50,7 @@ public class Main {
         double value;
         while (true) {
             System.out.print(prompt);
-            value = scanner.nextFloat();
+            value = scanner.nextDouble();
             if (value >= min && value <= max)
                 break;
             System.out.println("Please enter a value between " + min + " and " + max);
